python/requirements.txt
docker
PyYAML
jinja2
markupsafe
